# api/openapi.yaml
openapi: 3.0.3
info:
  title: sisyphus API
  description: Hybrid-search internal support assistant.
  version: 0.1.0
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Liveness/readiness probe.
      security: []
      responses:
        "200":
          description: Service health.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        default:
          $ref: "#/components/responses/Error"
  /search:
    post:
      operationId: search
      summary: Hybrid (FTS + vector) chunk search.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Ranked chunks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        default:
          $ref: "#/components/responses/Error"
  /context:
    post:
      operationId: context
      summary: Answer a support question from retrieved context (plan §14).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextRequest"
      responses:
        "200":
          description: Constructed answer with sources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextResponse"
        default:
          $ref: "#/components/responses/Error"
  /file:
    post:
      operationId: getFile
      summary: Retrieve actual file content from a source repository.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRequest"
      responses:
        "200":
          description: File content (full or line range).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        default:
          $ref: "#/components/responses/Error"
  /fetch:
    post:
      operationId: fetchURL
      summary: Fetch a URL from an operator-approved site.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchURLRequest"
      responses:
        "200":
          description: Fetch result (body + status + metadata).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchURLResponse"
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
        version:
          type: string
    Error:
      type: object
      required: [error_message]
      properties:
        error_message:
          type: string
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        service:
          type: string
        filters:
          type: object
          description: |
            Metadata filters applied as exact-match against chunk metadata.
            Well-known keys: status, source, jira_project, jira_component,
            jira_key, authority, repo. Values are always strings.
            When source is set here, source_tier/source_prefixes are ignored.
          additionalProperties:
            type: string
        source_tier:
          type: string
          description: |
            Source policy applied when filters.source is absent. Defaults to
            curated. Values: curated (docs, Jira, GitLab issues/MRs/releases,
            manifests), code (code and manifests), history (commits, tags,
            GitLab MRs/issues, Telegram), all (no source restriction).
        source_prefixes:
          type: array
          description: |
            Explicit source prefixes to search when filters.source is absent.
            Overrides source_tier. Prefixes match concrete per-repo sources such
            as git_docs:<repo>, git_code:<repo>, and git_commits:<repo>.
          items:
            type: string
        limit:
          type: integer
          format: int32
          default: 30
    SearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
    SearchResult:
      type: object
      required: [chunk_id, document_id, text, score]
      properties:
        chunk_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        source:
          type: string
        source_url:
          type: string
        title:
          type: string
        chunk_type:
          type: string
        text:
          type: string
        score:
          type: number
          format: double
        vector:
          type: boolean
    ContextRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
        service:
          type: string
        filters:
          type: object
          description: |
            Metadata filters applied as exact-match against chunk metadata.
            Well-known keys: status, source, jira_project, jira_component,
            jira_key, authority, repo. Values are always strings.
            When source is set here, source_tier/source_prefixes are ignored.
          additionalProperties:
            type: string
        source_tier:
          type: string
          description: |
            Source policy applied when filters.source is absent. Defaults to
            curated. Values: curated (docs, Jira, GitLab issues/MRs/releases,
            manifests), code (code and manifests), history (commits, tags,
            GitLab MRs/issues, Telegram), all (no source restriction).
        source_prefixes:
          type: array
          description: |
            Explicit source prefixes to search when filters.source is absent.
            Overrides source_tier. Prefixes match concrete per-repo sources such
            as git_docs:<repo>, git_code:<repo>, and git_commits:<repo>.
          items:
            type: string
    ContextResponse:
      type: object
      required: [answer, results]
      properties:
        answer:
          type: string
        confidence:
          type: string
        buttons:
          type: array
          description: |
            Actionable links surfaced alongside the answer (e.g. the sources it
            relied on), suitable for rendering as buttons. URLs are always
            absolute http(s).
          items:
            $ref: "#/components/schemas/Link"
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
    Link:
      type: object
      required: [text, url]
      properties:
        text:
          type: string
          description: Human-readable button label.
        url:
          type: string
          description: Absolute http(s) URL.
    FileRequest:
      type: object
      required: [repo, path]
      properties:
        repo:
          type: string
          description: Repository name as recorded in chunk metadata.
        path:
          type: string
          description: Repo-relative file path as recorded in chunk metadata.
        branch:
          type: string
          description: Optional branch (recorded in metadata but not used for lookup).
        start:
          type: integer
          format: int32
          description: Optional 1-indexed start line (inclusive).
        end:
          type: integer
          format: int32
          description: Optional 1-indexed end line (inclusive).
    FileResponse:
      type: object
      required: [content, found]
      properties:
        content:
          type: string
        source:
          type: string
          description: Where the content came from ("local_clone" or "database").
        found:
          type: boolean
    FetchURLRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
        method:
          type: string
          default: GET
        body:
          type: string
          description: Request body for POST/PUT/PATCH (optional).
        headers:
          type: object
          additionalProperties:
            type: string
          description: Additional request headers (optional). Credential headers are set by the server and cannot be overridden.
    FetchURLResponse:
      type: object
      required: [status_code, body, from_site]
      properties:
        status_code:
          type: integer
        body:
          type: string
        from_site:
          type: string
          description: Name of the allowlist site that matched.
        truncated:
          type: boolean
        headers:
          type: object
          additionalProperties:
            type: string